Move more functions over to private trying to make compilation consume less memory
